+++ ChangeLog.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Filesystem::rename()
 - Filesystem::truncate()
 - Filesystem::create()
+- Filesystem::flush()
 
 ## [0.0.3] - 2024-10-25